fix: resolved problem with types in db and repositories

+++ src/main/java/com/bsuir/newPortalBack/entities/UserNotificationId.java
@@ -6,6 +6,6 @@ import java.io.Serializable;
 
 @Embeddable
 public class UserNotificationId implements Serializable {
-  private Long userId;
-  private Long notificationId;
+  private int userId;
+  private int notificationId;
 }
